mk: Improve STRIP_DEBUG.
The first issue was that on platforms where strip was set to a noop the
command would succeed but the subsequent mv would fail.  Improve this by
actually checking for the output file.

The second issue is that not all strip commands support -g, so we need
an explicit gstrip tool for GNU strip (or equivalents, the FreeBSD
"elftoolchain" supports -g too).

While here simplify and tidy the install-strip-debug target, as well as
improve performance by skipping symlinks.